+++ source/extensions/filters/http/fault/fault_filter.h
@@ -55,7 +55,7 @@ public:
 private:
   uint64_t abort_percent_{};                                // 0-100
   uint64_t http_status_{};                                  // HTTP or gRPC return codes
-  envoy::type::FractionalPercent fixed_delay_percentage_{}; // 0-100
+  envoy::type::FractionalPercent fixed_delay_percentage_{}; // 0.0-100.0
   uint64_t fixed_duration_ms_{};                            // in milliseconds
   std::string upstream_cluster_; // restrict faults to specific upstream cluster
   std::vector<Http::HeaderUtility::HeaderData> fault_filter_headers_;